+++ src/vue/forms/PayForm.vue
@@ -53,6 +53,12 @@ import { ATOMIC_SWAP_BID_TYPES } from '@/js/const/atomic-swap-bid-types.const'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { vuexTypes } from '@/vuex'
 import { mapGetters } from 'vuex'
+import { Bus } from '@/js/helpers/event-bus'
+import { api } from '@/api'
+
+const EVENTS = {
+  reloadAtomicSwap: 'reload-atomic-swap',
+}
 
 export default {
   name: 'pay-form',
@@ -121,10 +127,18 @@ export default {
           this.form.promoCode,
           this.form.email
         )
-        if (atomicSwapBid.type === ATOMIC_SWAP_BID_TYPES.redirect) {
-          window.location.href = atomicSwapBid.payUrl
-        } else {
-          this.atomicSwapBidDetails = atomicSwapBid
+        switch (atomicSwapBid.type) {
+          case ATOMIC_SWAP_BID_TYPES.redirect:
+            window.location.href = atomicSwapBid.payUrl
+            break
+          case ATOMIC_SWAP_BID_TYPES.cryptoInvoice:
+            this.atomicSwapBidDetails = atomicSwapBid
+            break
+          case ATOMIC_SWAP_BID_TYPES.internal:
+            await api.signAndSendTransaction(atomicSwapBid.tx)
+            Bus.success('pay-form.success-msg')
+            this.$emit(EVENTS.reloadAtomicSwap)
+            break
         }
       } catch (e) {
         ErrorHandler.process(e)